+++ packages/utils/src/index.ts
@@ -1,2 +1,1 @@
 export * from './iframeEventHandler'
-export * from './iframeEventHandler/iframeEventHandler.types'
